Change permissions on pppd-ril
On devices that use a pppd for cellular network access, a pppd-ril
process needs to run as root to configure the connection.  Changing
the process to suid-root in the build image is the only known way
to avoid needing to remount the system image at boot to accomplish
the same thing.

Change-Id: I6c778391b8040ee818343e304891e7d341cc729b