Add new raw image sensor HAL pixel format.
This format is exposed outside of the HAL to applications.

RAW_SENSOR is a single-channel 16-bit format, typically representing raw
Bayer-pattern images from an image sensor, with minimal processing.

The exact pixel layout of the data in the buffer is sensor-dependent, and
needs to be queried from the camera device.

Generally, not all 16 bits are used; more common values are 10 or 12
bits. All parameters to interpret the raw data (black and white points,
color space, etc) must be queried from the camera device.

This format assumes
- an even width
- an even height
- a horizontal stride multiple of 16 pixels (32 bytes).

Change-Id: I67fc72d1c2c899e258fcd53b3ab91b459e9210f0