Raise the max file open limit in sdcard
The default is 1024 files, and in some testing, the limit has been
hit.  This raises the limit to 8192.  Going higher starts to cause
performance issues (I started to notice that around 16K open files
in my testing) as sdcard does linear searches.  If a higher max
is needed, then the sdcard daemon will need some optimizations.

Bug: 7442187

Change-Id: I7aba7f4556ed70651f36244294a6756f3d6b8963